feat: Update localization support for filters in resources page
This commit updates the localization support for the filters in the resources page. The "category" and "subCategory" labels in the Arabic and English JSON files have been modified to include a colon (:) at the end. This change ensures consistency and improves readability in the user interface.

The commit also includes updates to the package.json and pnpm-lock.yaml files. The "@radix-ui/react-popover" package has been added as a dependency with version 1.1.1. Additionally, the npm and TypeScript type dependencies have been updated.

Lastly, the commit introduces a new Popover component in the resources page. This component is used to display additional information when a row is clicked in the data table. The Popover component is implemented in the new "src/components/ui/popover.tsx" file, and it is used in the "src/app/[locale]/(dashboard)/resources/page.tsx" file.